requirements.txt
python-telegram-bot
pydantic
pydantic_settings
gallery-dl